+++ client/src/components/Login/Login.jsx
@@ -1,0 +1,99 @@
+import React, { useEffect, useState } from "react";
+import { useFormik } from "formik";
+import * as yup from "yup";
+
+export const LoginForm = () => {
+  const [users, setUsers] = useState([]);
+  const [refreshPage, setRefreshPage] = useState(false);
+  // Pass the useFormik() hook initial form values and a submit function that will
+  // be called when the form is submitted
+
+  useEffect(() => {
+    console.log("FETCH!");
+    fetch("/users")
+      .then((res) => res.json())
+      .then((data) => {
+        setUsers(data);
+        console.log(data);
+      });
+  }, [refreshPage]);
+
+  const formSchema = yup.object().shape({
+    email: yup
+      .string()
+      .email("invalid email")
+      .required("Must enter an email address"),
+    password: yup
+      .string()
+      .required("Password is required")
+      .min(6, "Password must be at least 6 characters"),
+  });
+
+  const formik = useFormik({
+    initialValues: {
+      email: "",
+      password: "",
+    },
+    validationSchema: formSchema,
+    onSubmit: (values) => {
+      fetch("users", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(values, null, 2),
+      }).then((res) => {
+        if (res.status == 200) {
+          setRefreshPage(!refreshPage);
+        }
+      });
+    },
+  });
+
+  return (
+    <div>
+      <form onSubmit={formik.handleSubmit} style={{ margin: "30px" }}>
+        <label htmlFor="email">Email Address</label>
+        <br />
+        <input
+          id="email"
+          name="email"
+          type="email"
+          onChange={formik.handleChange}
+          value={formik.values.email}
+        />
+        <p style={{ color: "red " }}>{formik.errors.email}</p>
+
+        <label htmlFor="password">Password</label>
+        <br />
+        <input
+          id="password"
+          name="password"
+          type="password"
+          onChange={formik.handleChange}
+          value={formik.values.password}
+        />
+        <p style={{ color: "red " }}>{formik.errors.password}</p>
+
+        <button type="submit">Login</button>
+      </form>
+
+      <table style={{ padding: "15px" }}>
+        <tbody>
+          <tr>
+            <th>Email</th>
+          </tr>
+          {users === "undefined" ? (
+            <p>Loading</p>
+          ) : (
+            users.map((user, i) => (
+              <tr key={i}>
+                <td>{user.emali}</td>
+              </tr>
+            ))
+          )}
+        </tbody>
+      </table>
+    </div>
+  );
+};
